changes proposed by reviewer:
. Added on and off callbacks for font shader control.
. Moved new classes and functions to their own file. Under imhexlib in helpers with the name of freetype so that it can be expanded if needed in the future.
. Removed old and unused code.

Additionally, the following changes are also included:
. Fixed problem with fonts becoming transparent and leaking the background colors. The equation  being used to blend alphas was using the wrong parameter of GL_SRC_ALPHA which would then multiply by itself creating additional transparencies. Using GL_ONE instead prevents the error from occurring.
. Tweaked the Gamma values to obtain better contrast between active and disabled items.
. Made some of the descriptions on the settings short enough so that they fit with default window size making sure te description was complete.
. Fixed logic inconsistency when turning antialiasing on in font configuration.
. Added drop down option in font settings to select among three types of antialiasing.
