+++ lib/libimhex/CMakeLists.txt
@@ -40,6 +40,7 @@ set(LIBIMHEX_SOURCES
         source/helpers/imgui_hooks.cpp
         source/helpers/semantic_version.cpp
         source/helpers/keys.cpp
+        source/helpers/freetype.cpp
 
         source/test/tests.cpp
