+++ plugins/builtin/romfs/lang/en_US.json
@@ -474,10 +474,10 @@
         "hex.builtin.setting.general.auto_backup_time.format.extended": "Every {0}m {1}s",
         "hex.builtin.setting.general.auto_load_patterns": "Auto-load supported pattern",
         "hex.builtin.setting.general.server_contact": "Enable update checks and usage statistics",
-        "hex.builtin.setting.general.max_mem_file_size": "Maximum size of file to load into memory",
+        "hex.builtin.setting.general.max_mem_file_size": "Max file size to load into RAM",
         "hex.builtin.setting.general.max_mem_file_size.desc": "Small files are loaded into memory to prevent them from being modified directly on disk.\n\nIncreasing this size allows larger files to be loaded into memory before ImHex resorts to streaming in data from disk.",
         "hex.builtin.setting.general.network_interface": "Enable network interface",
-        "hex.builtin.setting.general.pattern_data_max_filter_items": "Max number of items shown when filtering pattern data",
+        "hex.builtin.setting.general.pattern_data_max_filter_items": "Max filtered pattern items shown",
         "hex.builtin.setting.general.save_recent_providers": "Save recently used providers",
         "hex.builtin.setting.general.show_tips": "Show tips on startup",
         "hex.builtin.setting.general.sync_pattern_source": "Sync pattern source code between providers",
