+++ plugins/fonts/include/font_settings.hpp
@@ -3,6 +3,14 @@
 #include <hex/api/content_registry.hpp>
 
 namespace hex::fonts {
+    class AntialiasPicker : public ContentRegistry::Settings::Widgets::DropDown {
+    public:
+        AntialiasPicker() : DropDown(
+                std::vector<UnlocalizedString>({"hex.fonts.setting.font.antialias_none", "hex.fonts.setting.font.antialias_grayscale", "hex.fonts.setting.font.antialias_subpixel"}),
+                std::vector<nlohmann::json>({"none" , "grayscale" , "subpixel"}),
+                nlohmann::json({"subpixel"})
+                ){}
+    };
 
     class FontFilePicker : public ContentRegistry::Settings::Widgets::FilePicker {
     public:
@@ -31,7 +39,7 @@ namespace hex::fonts {
 
     class FontSelector : public ContentRegistry::Settings::Widgets::Widget {
     public:
-        FontSelector() : m_fontSize(16, 2, 100), m_bold(false), m_italic(false), m_antiAliased(true) { }
+        FontSelector() : m_fontSize(16, 2, 100), m_antiAliased(), m_bold(false), m_italic(false) { }
 
         bool draw(const std::string &name) override;
 
@@ -43,7 +51,7 @@ namespace hex::fonts {
         [[nodiscard]] float getFontSize() const;
         [[nodiscard]] bool isBold() const;
         [[nodiscard]] bool isItalic() const;
-        [[nodiscard]] bool isAntiAliased() const;
+        [[nodiscard]] const std::string antiAliasingType() const;
 
     private:
         bool drawPopup();
@@ -51,7 +59,8 @@ namespace hex::fonts {
     private:
         FontFilePicker m_fontFilePicker;
         SliderPoints m_fontSize;
-        ContentRegistry::Settings::Widgets::Checkbox m_bold, m_italic, m_antiAliased;
+        AntialiasPicker m_antiAliased;
+        ContentRegistry::Settings::Widgets::Checkbox m_bold, m_italic;
 
         bool m_applyEnabled = false;
     };
